[FIX] - Validation on register items

+++ app/src/main/java/com/mentalhealth/eifie/data/models/response/UserPatientResponse.kt
@@ -2,6 +2,6 @@ package com.mentalhealth.eifie.data.models.response
 
 data class UserPatientResponse(
     val patientId: Long? = null,
-    val psychologistAssignedId: Int? = null,
+    val psychologistAssignedId: Long? = null,
     val user: UserResponse? = null
 )
